+++ Mips-net/Module/IRfDriverModule.cs
@@ -6,12 +6,12 @@ namespace Mips.Module
 {
     public interface IRfDriverModule
     {
-        Task<Unit> SetFrequency(string channel, double frequencyInHz);
+        Task<Unit> SetFrequency(string channel, int frequencyInHz);
         Task<Unit> SetRfPeakToPeak(string channel,double peakToPeakVoltage);
        
         Task<Unit> SetDriveLevel(string channel, double drive);
 
-        Task<double> GetFrequency(string channel);
+        Task<int> GetFrequency(string channel);
         Task<double> GetRFPositive(string channel);
         Task<double> GetRFNegative(string channel);
         Task<double> GetOutputDriveLevelPercent(string channel);
